kmir/{options,__main__}: move options classes into own module
