+++ kmir/src/kmir/options.py
@@ -1,0 +1,107 @@
+from __future__ import annotations
+
+import logging
+from dataclasses import dataclass
+from pathlib import Path
+from typing import TYPE_CHECKING
+
+from pyk.cli.args import LoggingOptions
+
+if TYPE_CHECKING:
+    from typing import Final
+
+_LOGGER: Final = logging.getLogger(__name__)
+_LOG_FORMAT: Final = '%(levelname)s %(asctime)s %(name)s - %(message)s'
+
+
+@dataclass
+class KMirOpts(LoggingOptions): ...
+
+
+@dataclass
+class RunOpts(KMirOpts):
+    bin: str | None
+    file: str | None
+    depth: int
+    start_symbol: str
+    haskell_backend: bool
+
+
+@dataclass
+class GenSpecOpts(KMirOpts):
+    input_file: Path
+    output_file: Path | None
+    start_symbol: str
+
+    def __init__(self, input_file: Path, output_file: Path | str | None, start_symbol: str) -> None:
+        self.input_file = input_file
+        if output_file is None:
+            self.output_file = None
+        else:
+            self.output_file = Path(output_file).resolve()
+        self.start_symbol = start_symbol
+
+
+@dataclass
+class ProveRSOpts(KMirOpts):
+    rs_file: Path
+    bug_report: Path | None
+    max_depth: int | None
+    max_iterations: int | None
+
+    def __init__(
+        self,
+        rs_file: Path,
+        bug_report: Path | None = None,
+        max_depth: int | None = None,
+        max_iterations: int | None = None,
+    ) -> None:
+        self.rs_file = rs_file
+        self.bug_report = bug_report
+        self.max_depth = max_depth
+        self.max_iterations = max_iterations
+
+
+@dataclass
+class ProveRunOpts(KMirOpts):
+    spec_file: Path
+    proof_dir: Path | None
+    include_labels: tuple[str, ...] | None
+    exclude_labels: tuple[str, ...] | None
+    bug_report: Path | None
+    max_depth: int | None
+    max_iterations: int | None
+    reload: bool
+
+    def __init__(
+        self,
+        spec_file: Path,
+        proof_dir: Path | str | None,
+        include_labels: str | None,
+        exclude_labels: str | None,
+        bug_report: Path | None = None,
+        max_depth: int | None = None,
+        max_iterations: int | None = None,
+        reload: bool = False,
+    ) -> None:
+        self.spec_file = spec_file
+        self.proof_dir = Path(proof_dir).resolve() if proof_dir is not None else None
+        self.include_labels = tuple(include_labels.split(',')) if include_labels is not None else None
+        self.exclude_labels = tuple(exclude_labels.split(',')) if exclude_labels is not None else None
+        self.bug_report = bug_report
+        self.max_depth = max_depth
+        self.max_iterations = max_iterations
+        self.reload = reload
+
+
+@dataclass
+class ProveViewOpts(KMirOpts):
+    proof_dir: Path
+    id: str
+
+
+@dataclass
+class ProvePruneOpts(KMirOpts):
+    proof_dir: Path
+    id: str
+    node_id: int
